refactor: modularized factories in main layer
