+++ src/main/factories/controllers/login/login-validation-factory.ts
@@ -1,6 +1,6 @@
-import { EmailValidation, ValidatorComposite, RequiredFieldValidation } from '../../../presentation/helpers/validators'
-import { Validator } from '../../../presentation/protocols'
-import { EmailValidatorAdapter } from '../../adapters/validators/email-validator-adapter'
+import { EmailValidation, ValidatorComposite, RequiredFieldValidation } from '../../../../presentation/helpers/validators'
+import { Validator } from '../../../../presentation/protocols'
+import { EmailValidatorAdapter } from '../../../adapters/validators/email-validator-adapter'
 
 const makeValidations = (): Validator[] => {
   const validations: Validator[] = []
