+++ src/main/factories/decorators/log-controller-decorator-factory.ts
@@ -1,0 +1,9 @@
+import { LogMongoRepository } from '../../../infra/db/mongodb/log/log-mongo-repository'
+import { Controller } from '../../../presentation/protocols'
+import { LoggerControllerDecorator } from '../../decorators/log-controller-decorator'
+
+export const makeLogControllerDecorator = (controller: Controller): Controller => {
+  const logMongoRepository = new LogMongoRepository()
+
+  return new LoggerControllerDecorator(controller, logMongoRepository)
+}
